Added gunshot sound effects.
TODO: Sometimes you will attempt to fire before the fire rate allows you to.
This feels very clunky a lot of the time.

It would be nice to queue up a shot if you attempt to fire within
a certain threshold, and have it fire as soon as possible.

This idea is similar to the reasoning behind "coyote time" for platformers.