Created an item pickup system.
The system is naive in a sense. If there are multiple items to pick up
that are very near each other, one will probably dominate the stack
rendering the player unable to pick up the others. This shouldn't be an
issue for this game in particular, and won't be fixed unless need be.
It would cause the system to become much more complex.